Rename Show.jsx to match function name.
