+++ app/frontend/components/AdminEnrollmentShow.jsx
