+++ app/frontend/entrypoints/admin_enrollment_show.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import AdminEnrollmentShow from '~/components/Show';
+import AdminEnrollmentShow from '~/components/AdminEnrollmentShow';
 
 const container = document.getElementById('show');
 const root = ReactDOM.createRoot(container);
